requirements.txt
Django==4.1
djangorestframework==3.14.0
Pillow==9.3.0
qrcode==7.4.2
requests==2.28.2
